# perf(mpsc): rewrite and optimize wait queue