--- a/src/util/mutex.rs
+++ b/src/util/mutex.rs
@@ -1,5 +1,5 @@
 feature! {
-    #![feature = "std"]
+    #![all(feature = "std", not(feature = "parking_lot"))]
     pub(crate) use self::std_impl::*;
     mod std_impl;
 }
@@ -9,3 +9,9 @@ pub(crate) use self::spin_impl::*;
 
 #[cfg(any(not(feature = "std"), test))]
 mod spin_impl;
+
+feature! {
+    #![all(feature = "std", feature = "parking_lot")]
+    #[allow(unused_imports)]
+    pub(crate) use parking_lot::{Mutex, MutexGuard};
+}